refactor: 💡 Tests refactored to use contract upgrade
