+++ utils/index.ts
@@ -1,18 +1,53 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { utils, BigNumber, TypedDataField, VoidSigner } from "ethers";
-import {
-  PAYMENT_OPTION_TYPE_HASH,
-  CANCEL_OPTION_TYPE_HASH,
-  OFFER_TYPE_HASH,
-} from "./constants";
-import { MockERC20Dec18Permit } from "../typechain";
-import {
-  PaymentOption,
-  CancelOption,
-  OfferPayload,
-  Offer,
-  Request,
-} from "./types";
+import { utils, BigNumber } from "ethers";
+
+// Protocol entities types (kinds) as object
+export const kinds = {
+  supplier: utils.formatBytes32String("supplier"),
+  retailer: utils.formatBytes32String("retailer"),
+};
+
+// Protocol entities types (kinds) as values array
+export const kindsArr = Object.values(kinds);
+
+// Protocol defaults
+export const eip712name = "Market";
+export const eip712version = "1";
+export const minDeposit = BigNumber.from("1000000000000000000000");
+export const claimPeriod = BigNumber.from("60");
+export const protocolFee = BigNumber.from("1");
+export const retailerFee = BigNumber.from("1");
+
+// ethers.solidityPackedKeccak256(
+//   ['string'],
+//   ['PaymentOption(bytes32 id,uint256 price,address asset)'],
+// );
+export const PAYMENT_OPTION_TYPE_HASH =
+  "0x2f8fc0b3ad3f58f6deb367673d38e4112a3c8c64de033c5b780b84ef8f67cde6";
+
+// ethers.solidityPackedKeccak256(
+//   ['string'],
+//   ['CancelOption(uint256 time,uint256 penalty)'],
+// );
+export const CANCEL_OPTION_TYPE_HASH =
+  "0x8ea27057ea8a0239f02c8b75748218a035a5a2a2a0785b53aaa99af91ff538c5";
+
+// ethers.solidityPackedKeccak256(
+//   ['string'],
+//   [
+//     'Offer(bytes32 id,uint256 expire,bytes32 supplierId,uint256 chainId,bytes32 requestHash,bytes32 optionsHash,bytes32 paymentHash,bytes32 cancelHash,bool transferable,uint256 checkIn)',
+//   ],
+// );
+export const OFFER_TYPE_HASH =
+  "0x4fb12343a6f44152999c71291770d97fc1eace9d7d04889330d5a6d1af4a57c7";
+
+// ethers.solidityPackedKeccak256(
+//   ['string'],
+//   [
+//     'Voucher(bytes32 id,address signer)',
+//   ],
+// );
+export const CHECK_IN_TYPE_HASH =
+  "0xf811d7f3ddb148410001929e2cbfb7fea8779b9349b7c2f650fa91840528d69c";
 
 export const randomId = (): string =>
   utils.solidityKeccak256(
@@ -30,282 +65,3 @@ export const randomId = (): string =>
 
 export const createSupplierId = (address: string, salt: string): string =>
   utils.solidityKeccak256(["address", "bytes32"], [address, salt]);
-
-export const hashObject = (request: unknown): string =>
-  utils.solidityKeccak256(["string"], [JSON.stringify(request)]);
-
-export const hashPaymentOption = (option: PaymentOption): string =>
-  utils.solidityKeccak256(
-    ["bytes32", "bytes32", "uint256", "address"],
-    [PAYMENT_OPTION_TYPE_HASH, option.id, option.price, option.asset]
-  );
-
-export const hashCancelOption = (option: CancelOption): string =>
-  utils.solidityKeccak256(
-    ["bytes32", "uint256", "uint256"],
-    [CANCEL_OPTION_TYPE_HASH, option.time, option.penalty]
-  );
-
-export const hashPaymentOptionArray = (options: PaymentOption[]): string => {
-  const hashes = [];
-
-  for (let i = 0; i < options.length; i++) {
-    hashes[i] = hashPaymentOption(options[i]);
-  }
-  return utils.solidityKeccak256(["bytes32[]"], [hashes]);
-};
-
-export const hashCancelOptionArray = (options: CancelOption[]): string => {
-  const hashes = [];
-
-  for (let i = 0; i < options.length; i++) {
-    hashes[i] = hashCancelOption(options[i]);
-  }
-
-  return utils.solidityKeccak256(["bytes32[]"], [hashes]);
-};
-
-export const hashOfferPayload = (payload: OfferPayload): string =>
-  utils.solidityKeccak256(
-    [
-      "bytes32",
-      "bytes32",
-      "uint256",
-      "bytes32",
-      "uint256",
-      "bytes32",
-      "bytes32",
-      "bytes32",
-      "bytes32",
-      "bool",
-      "uint256",
-    ],
-    [
-      OFFER_TYPE_HASH,
-      payload.id,
-      payload.expire,
-      payload.supplierId,
-      payload.chainId,
-      payload.requestHash,
-      payload.optionsHash,
-      payload.paymentHash,
-      payload.cancelHash,
-      payload.transferable,
-      payload.checkIn,
-    ]
-  );
-
-export const hashCheckInOut = (offerId: string, signer: string): string =>
-  utils.solidityKeccak256(
-    ["bytes32", "bytes32", "address"],
-    [OFFER_TYPE_HASH, offerId, signer]
-  );
-
-export const offerEip712Types: Record<string, Array<TypedDataField>> = {
-  Offer: [
-    {
-      name: "id",
-      type: "bytes32",
-    },
-    {
-      name: "expire",
-      type: "uint256",
-    },
-    {
-      name: "supplierId",
-      type: "bytes32",
-    },
-    {
-      name: "chainId",
-      type: "uint256",
-    },
-    {
-      name: "requestHash",
-      type: "bytes32",
-    },
-    {
-      name: "optionsHash",
-      type: "bytes32",
-    },
-    {
-      name: "paymentHash",
-      type: "bytes32",
-    },
-    {
-      name: "cancelHash",
-      type: "bytes32",
-    },
-    {
-      name: "transferable",
-      type: "bool",
-    },
-    {
-      name: "checkIn",
-      type: "uint256",
-    },
-  ],
-};
-
-export const checkInOutTypes: Record<string, Array<TypedDataField>> = {
-  Voucher: [
-    {
-      name: "id",
-      type: "bytes32",
-    },
-    {
-      name: "signer",
-      type: "address",
-    },
-  ],
-};
-
-export const createCheckInOutSignature = async (
-  signer: VoidSigner,
-  offerId: string,
-  name: string,
-  version: string,
-  chainId: BigNumber,
-  verifyingContract: string
-): Promise<string> =>
-  await signer._signTypedData(
-    {
-      name,
-      version,
-      chainId,
-      verifyingContract,
-    },
-    checkInOutTypes,
-    {
-      id: offerId,
-      signer: signer.address,
-    }
-  );
-
-export const createPermitSignature = async (
-  signer: VoidSigner,
-  erc20: MockERC20Dec18Permit,
-  owner: string,
-  spender: string,
-  value: BigNumber,
-  deadline: number,
-  version = "1"
-): Promise<string> => {
-  const nonce = await erc20.nonces(owner);
-  const name = await erc20.name();
-  const chainId = await signer.getChainId();
-
-  return await signer._signTypedData(
-    {
-      name,
-      version,
-      chainId,
-      verifyingContract: erc20.address,
-    },
-    {
-      Permit: [
-        {
-          name: "owner",
-          type: "address",
-        },
-        {
-          name: "spender",
-          type: "address",
-        },
-        {
-          name: "value",
-          type: "uint256",
-        },
-        {
-          name: "nonce",
-          type: "uint256",
-        },
-        {
-          name: "deadline",
-          type: "uint256",
-        },
-      ],
-    },
-    {
-      owner,
-      spender,
-      value,
-      nonce,
-      deadline,
-    }
-  );
-};
-
-export const getCancelPenalty = (
-  options: CancelOption[],
-  timestamp: BigNumber
-) => {
-  let selectedTime = BigNumber.from(0);
-  let selectedPenalty = BigNumber.from(0);
-
-  for (const option of options) {
-    if (
-      timestamp.gte(option.time) &&
-      (selectedTime.isZero() || option.time.lt(selectedTime))
-    ) {
-      selectedTime = option.time;
-      selectedPenalty = option.penalty;
-    }
-  }
-
-  return selectedPenalty.lte(BigNumber.from(100))
-    ? selectedPenalty
-    : BigNumber.from(100);
-};
-
-export const buildOffer = async (
-  signer: VoidSigner,
-  supplierId: string,
-  expire: BigNumber,
-  checkIn: BigNumber,
-  checkOut: BigNumber,
-  request: Request,
-  offerOptions: object,
-  payment: PaymentOption[],
-  cancel: CancelOption[],
-  transferableOverride: boolean,
-  name: string,
-  version: string,
-  chainId: BigNumber,
-  verifyingContract: string
-): Promise<Offer> => {
-  const offerPayload: OfferPayload = {
-    id: randomId(),
-    expire,
-    supplierId: supplierId,
-    chainId,
-    requestHash: hashObject(request),
-    optionsHash: hashObject(offerOptions),
-    paymentHash: hashPaymentOptionArray(payment),
-    cancelHash: hashCancelOptionArray(cancel),
-    transferable: transferableOverride ?? Math.random() > 0.5,
-    checkIn,
-    checkOut,
-  };
-
-  const signature = await signer._signTypedData(
-    {
-      name,
-      version,
-      chainId,
-      verifyingContract,
-    },
-    offerEip712Types,
-    offerPayload
-  );
-
-  const offer: Offer = {
-    request,
-    options: offerOptions,
-    payment,
-    cancel,
-    payload: offerPayload,
-    signature,
-  };
-
-  return offer;
-};
